Don't set information not needed on enum signals
When creating a PWire for a enum type the signedness as well as whether
the base type is an integer is assigned to the wire.

But this information is never queried again. When creating the netenum_t
this information is directly taken from the enum_type_t.

The signedness and integer information of the PWire is only used when
elaborating a netvector_t.

Removing this makes the pfrom_set_enum() function similar to those for
other types and will allow us to consolidate them in follow up patches.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>